Tests/LibWeb/Text/expected/HTML/opaque-origin-same-origin-equality.txt

Can't access property 'thing' on cross-origin object
